# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Tameshare MVP API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /v1/practices:
    get:
      summary: List practices newest first
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Practice page
          content: {application/json: {schema: {$ref: '#/components/schemas/PracticePage'}}}
        '422': {$ref: '#/components/responses/ValidationError'}
  /v1/practices/{practiceId}:
    get:
      summary: Get a practice with its sources
      parameters: [{$ref: '#/components/parameters/PracticeId'}]
      responses:
        '200': {description: Practice detail, content: {application/json: {schema: {$ref: '#/components/schemas/PracticeDetail'}}}}
        '400': {$ref: '#/components/responses/InvalidId'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/practices/{practiceId}/experiences:
    get:
      summary: List experiences for a practice
      parameters: [{$ref: '#/components/parameters/PracticeId'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '200': {description: Experience page, content: {application/json: {schema: {$ref: '#/components/schemas/ExperiencePage'}}}}
        '400': {$ref: '#/components/responses/InvalidId'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
    post:
      summary: Record an experience for the authenticated user
      security: [{bearerUserId: []}]
      parameters: [{$ref: '#/components/parameters/PracticeId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExperienceInput'}}}}
      responses:
        '201': {description: Created experience, content: {application/json: {schema: {$ref: '#/components/schemas/Experience'}}}}
        '400': {$ref: '#/components/responses/InvalidId'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /v1/experiences/{experienceId}:
    patch:
      summary: Update the authenticated user's experience
      security: [{bearerUserId: []}]
      parameters:
        - name: experienceId
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExperienceInput'}}}}
      responses:
        '200': {description: Updated experience, content: {application/json: {schema: {$ref: '#/components/schemas/Experience'}}}}
        '400': {$ref: '#/components/responses/InvalidId'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/ValidationError'}
  /v1/listing-requests:
    post:
      summary: Submit a source URL listing request
      description: Returns 201 with created=true for a new request, or 200 with created=false and the existing request for a duplicate URL.
      security: [{bearerUserId: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ListingRequestInput'}}}}
      responses:
        '200': {description: Existing request, content: {application/json: {schema: {$ref: '#/components/schemas/ListingRequestResult'}}}}
        '201': {description: New request, content: {application/json: {schema: {$ref: '#/components/schemas/ListingRequestResult'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '422': {$ref: '#/components/responses/ValidationError'}
components:
  securitySchemes:
    bearerUserId:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: Existing anonymous-auth user UUID. The client cannot choose another user in the JSON body.
  parameters:
    PracticeId: {name: practiceId, in: path, required: true, schema: {type: string, format: uuid}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100, default: 20}}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
  responses:
    InvalidId: {description: Invalid UUID, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Unauthorized: {description: Missing or invalid authentication, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Forbidden: {description: The experience belongs to another user, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    NotFound: {description: Resource not found, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    ValidationError: {description: Domain or pagination validation failed, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Practice: {type: object, required: [id, title, created_at], properties: {id: {type: string, format: uuid}, title: {type: string}, created_at: {type: string, format: date-time}}}
    Source: {type: object, required: [url], properties: {url: {type: string, format: uri}}}
    PracticeDetail: {allOf: [{$ref: '#/components/schemas/Practice'}, {type: object, required: [sources], properties: {sources: {type: array, items: {$ref: '#/components/schemas/Source'}}}}]}
    PracticePage: {type: object, required: [items, limit, offset, has_more], properties: {items: {type: array, items: {$ref: '#/components/schemas/Practice'}}, limit: {type: integer}, offset: {type: integer}, has_more: {type: boolean}}}
    ExperienceInput: {type: object, properties: {note: {type: string, maxLength: 100, nullable: true}}, additionalProperties: false}
    Experience: {type: object, required: [id, practice_id, user_id, note, created_at, updated_at], properties: {id: {type: string, format: uuid}, practice_id: {type: string, format: uuid}, user_id: {type: string, format: uuid}, note: {type: string, nullable: true}, created_at: {type: string, format: date-time}, updated_at: {type: string, format: date-time}}}
    ExperiencePage: {type: object, required: [items, limit, offset, has_more], properties: {items: {type: array, items: {$ref: '#/components/schemas/Experience'}}, limit: {type: integer}, offset: {type: integer}, has_more: {type: boolean}}}
    ListingRequestInput: {type: object, required: [source_url], properties: {source_url: {type: string, format: uri}}, additionalProperties: false}
    ListingRequestResult: {type: object, required: [id, status, created], properties: {id: {type: string, format: uuid}, status: {type: string, enum: [pending, accepted, rejected]}, created: {type: boolean}}}
    Error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}